requirements.txt
pandas==0.23.4
Django==2.2.4
nltk==3.3
scipy==1.2.1
python_Levenshtein==0.12.0
xlrd==1.1.0